+++ web/src/main/java/com/victory/hr/sys/dao/RoleDao.java
@@ -7,5 +7,5 @@ import com.victory.hr.sys.entity.User;
 /**
  * Created by ajkx on 2017/5/7.
  */
-public interface RoleDao extends BaseDao<Role,Long> {
+public interface RoleDao extends BaseDao<Role,Integer> {
 }
