+++ web/src/main/java/com/victory/hr/sys/dao/impl/UserDaoImpl.java
@@ -11,7 +11,7 @@ import java.util.List;
  * Created by ajkx on 2017/5/7.
  */
 @Repository
-public class UserDaoImpl extends BaseDaoImpl<User,Long> implements UserDao{
+public class UserDaoImpl extends BaseDaoImpl<User,Integer> implements UserDao{
     public User findByUserName(String username) {
         List<User> list = find("select u from User u where u.name = ?0", username);
         if (list.size() == 0) {
