+++ web/src/main/java/com/victory/hr/sys/service/RoleService.java
@@ -1,8 +1,7 @@
 package com.victory.hr.sys.service;
 
-import com.victory.hr.common.BaseService;
+import com.victory.hr.common.service.BaseService;
 import com.victory.hr.sys.entity.Role;
-import com.victory.hr.sys.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +9,7 @@ import org.springframework.stereotype.Service;
  * Created by ajkx on 2017/5/7.
  */
 @Service
-public class RoleService extends BaseService<Role,Long>{
+public class RoleService extends BaseService<Role,Integer>{
 
     @Autowired
     private ResourceService resourceService;
